PolicyGPT/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
openai>=1.13.0
python-dotenv==1.0.1
PyPDF2==3.0.1
azure-ai-translation-text==1.0.0
aiohttp>=3.9.0  # For async HTTP requests
cachetools>=5.3.0  # For caching support